+++ R/create_leaderdays.R
@@ -40,9 +40,27 @@ create_leaderdays <- function(system = "archigos", standardize_cow = TRUE) {
       unnest(.data$date) %>%
       mutate(year = .pshf_year(.data$date)) -> leaderdays
 
+    # I need to do some ad hoc corrections here
+    # There are pieces of the gw_cow_years data frame where G-W says the state doesn't exist but it appears in the CoW data.
+    # Because G-W is the "master" here, this is going to happen.
+    # The gw_cow_years data would be functionally correct here, but that's immaterial for matching it to the Archigos data.
+    # gw_cow_years %>% filter(is.na(gwcode) & year >= 1870) %>% group_by(stateabb) %>% summarize(ccode = max(ccode, na.rm=T), n = n(), min = min(year), max = max(year))
+    # For example: G-W have Morocco drop out at 1904. CoW has Morocco drop out at 1912.
+    # If you want to standardize to CoW, you're going to miss that.
+    # You'll also miss that G-W has Saudi Arabia start in 1932, but CoW has it at 1927.
+    # Archigos has leader data for the Saudis to 1927, so you'll want that.
+
+
     leaderdays %>%
       left_join(., gw_cow_years %>% select(.data$gwcode, .data$ccode, .data$year)) %>%
       select(.data$obsid, .data$gwcode, .data$ccode, everything()) -> leaderdays
+
+    leaderdays %>%
+      mutate(ccode = case_when(
+        .data$gwcode == 600 & (.data$year >= 1905 & .data$year <= 1912) ~ 600,
+        .data$gwcode == 670 & (.data$year >= 1927 & .data$year <= 1932) ~ 670,
+        TRUE ~ .data$ccode
+      )) -> leaderdays
 
     if (standardize_cow == TRUE) {