+++ R/whittle_conflicts_fatality.R
@@ -5,7 +5,7 @@
 #' of whittling dyadic dispute-year data into true dyad-year data (like in the Gibler-Miller-Little conflict data). This particular
 #' function will keep the observations with the highest observed fatality.
 #'
-#' @return \code{whittle_conflicts_fatality()} takes a dyad-year data frame with a declared conflict attribute type and, grouping by the
+#' @return \code{whittle_conflicts_fatality()} takes a dyad-year data frame or leader-dyad-year data frame with a declared conflict attribute type and, grouping by the
 #' dyad and year, returns just those observations that have the highest observed dispute-level fatality.
 #' This will not eliminate all duplicates, far from it, but it's a sensible second cut (after whittling onsets in
 #' \code{whittle_conflicts_onsets()} the extent to which dispute-level fatality is a good heuristic for dispute-level severity/importance.
@@ -109,7 +109,40 @@ whittle_conflicts_fatality <- function(data) {
 
 
 
-  } else  {
+  } else if (length(attributes(data)$ps_data_type) > 0 && attributes(data)$ps_data_type == "leader_dyad_year" &&  attributes(data)$ps_conflict_type == "gml") {
+
+    data[ , c('styear', 'stmon', 'settle', 'fatality', 'mindur', 'maxdur', 'hiact', 'hostlev', 'recip', 'outcome')] <- list(NULL)
+
+    attr_ps_data_type <- attributes(data)$ps_data_type
+    attr_ps_system <- attributes(data)$ps_system
+    attr_ps_conflict_type <- attributes(data)$ps_conflict_type
+
+    data %>%
+      left_join(., gml_mid_disps ) -> hold_this
+
+    hold_this %>%
+      mutate(fatality = ifelse(.data$fatality == -9, .5, .data$fatality)) %>%
+      arrange(.data$ccode1, .data$obsid1, .data$ccode2, .data$obsid2, .data$year) %>%
+      group_by(.data$ccode1, .data$obsid1, .data$ccode2, .data$obsid2, .data$year) %>%
+      mutate(duplicated = ifelse(n() > 1, 1, 0)) %>%
+      group_by(.data$ccode1, .data$ccode2, .data$year, .data$duplicated) %>%
+      # Keep the highest fatality
+      filter(.data$fatality == max(.data$fatality)) %>%
+      mutate(fatality = ifelse(.data$fatality == .5, -9, .data$fatality)) %>%
+      arrange(.data$ccode1, .data$obsid1, .data$ccode2, .data$obsid2, .data$year) %>%
+      # practice safe group_by()
+      ungroup() %>%
+      select(-.data$duplicated) -> data
+
+    data[ , c('styear', 'stmon', 'settle', 'fatality', 'mindur', 'maxdur', 'hiact', 'hostlev', 'recip', 'outcome')] <- list(NULL)
+
+    attr(data, "ps_data_type") <- attr_ps_data_type
+    attr(data, "ps_system") <-  attr_ps_system
+    attr(data, "ps_conflict_type") <-  attr_ps_conflict_type
+
+
+
+  } else {
     stop("whittle_conflicts_fatalities() doesn't recognize the data supplied to it.")
   }