+++ R/whittle_conflicts_onsets.R
@@ -5,7 +5,7 @@
 #' of whittling dyadic dispute-year data into true dyad-year data (like in the Gibler-Miller-Little conflict data). This particular
 #' function will drop ongoing conflicts in the presence of unique onsets.
 #'
-#' @return \code{whittle_conflicts_onsets()} takes a dyad-year data frame with a declared conflict attribute type and, grouping by the
+#' @return \code{whittle_conflicts_onsets()} takes a dyad-year data frame or leader-dyad-year data frame  with a declared conflict attribute type and, grouping by the
 #' dyad and year, returns just those observations with unique onsets where duplicates exist. This will not eliminate all duplicates, far
 #' from it, but it's a sensible place to start.
 #'
@@ -100,7 +100,43 @@ whittle_conflicts_onsets <- function(data) {
 
 
 
-  } else  {
+  } else if (length(attributes(data)$ps_data_type) > 0 && attributes(data)$ps_data_type == "leader_dyad_year" &&  attributes(data)$ps_conflict_type == "gml") {
+
+    # I want to err on the side of caution and do this merge every time, irregarding if the data are already there.
+    # It won't take much time.
+
+    data[ , c('styear', 'stmon', 'settle', 'fatality', 'mindur', 'maxdur', 'hiact', 'hostlev', 'recip', 'outcome')] <- list(NULL)
+
+    attr_ps_data_type <- attributes(data)$ps_data_type
+    attr_ps_system <- attributes(data)$ps_system
+    attr_ps_conflict_type <- attributes(data)$ps_conflict_type
+
+    data %>%
+      left_join(., gml_mid_disps ) -> hold_this
+
+
+    hold_this %>%
+      group_by(.data$ccode1, .data$obsid1, .data$ccode2, .data$obsid2, .data$year)  %>%
+      mutate(duplicated = ifelse(n() > 1, 1, 0)) %>%
+      # Remove anything that's not a unique MID onset
+      mutate(sddd = sd(.data$gmlmidonset),
+             sddd = ifelse(is.na(.data$sddd), 0, .data$sddd)) %>%
+      mutate(removeme = ifelse(.data$duplicated == 1 & .data$gmlmidonset == 0 & .data$sddd > 0, 1, 0)) %>%
+      filter(.data$removeme != 1) %>%
+      # remove detritus
+      select(-.data$removeme, -.data$sddd) %>%
+      # practice safe group_by()
+      ungroup() -> data
+
+    data[ , c('styear', 'stmon', 'settle', 'fatality', 'mindur', 'maxdur', 'hiact', 'hostlev', 'recip', 'outcome')] <- list(NULL)
+
+    attr(data, "ps_data_type") <- attr_ps_data_type
+    attr(data, "ps_system") <-  attr_ps_system
+    attr(data, "ps_conflict_type") <-  attr_ps_conflict_type
+
+
+
+  } else {
     stop("whittle_conflicts_onsets() doesn't recognize the data supplied to it.")
   }