+++ R/whittle_conflicts_startmonth.R
@@ -5,7 +5,7 @@
 #' of whittling dyadic dispute-year data into true dyad-year data (like in the Gibler-Miller-Little conflict data). This particular
 #' function will keep the observations that have the lowest start month.
 #'
-#' @return \code{whittle_conflicts_startmonth()} takes a dyad-year data frame with a declared conflict attribute type and, grouping by the
+#' @return \code{whittle_conflicts_startmonth()} takes a dyad-year data frame or leader-dyad-year data frame  with a declared conflict attribute type and, grouping by the
 #' dyad and year, returns just those observations that have the lowest start month.
 #'
 #' @details Dyads are capable of having multiple disputes in a given year, which can create a problem
@@ -109,6 +109,39 @@ whittle_conflicts_startmonth <- function(data) {
     attr(data, "ps_system") <-  attr_ps_system
     attr(data, "ps_conflict_type") <-  attr_ps_conflict_type
 
+
+
+  } else if (length(attributes(data)$ps_data_type) > 0 && attributes(data)$ps_data_type == "leader_dyad_year" &&  attributes(data)$ps_conflict_type == "gml") {
+
+    data[ , c('styear', 'stmon', 'settle', 'fatality', 'mindur', 'maxdur', 'hiact', 'hostlev', 'recip', 'outcome')] <- list(NULL)
+
+    attr_ps_data_type <- attributes(data)$ps_data_type
+    attr_ps_system <- attributes(data)$ps_system
+    attr_ps_conflict_type <- attributes(data)$ps_conflict_type
+
+    data %>%
+      left_join(., gml_mid_disps ) -> hold_this
+
+    hold_this %>%
+      arrange(.data$ccode1, .data$obsid1, .data$ccode2, .data$obsid2, .data$year) %>%
+      group_by(.data$ccode1, .data$obsid1, .data$ccode2, .data$obsid2, .data$year) %>%
+      mutate(duplicated = ifelse(n() > 1, 1, 0)) -> data
+
+    data %>%
+      group_by(.data$ccode1, .data$obsid1, .data$ccode2, .data$obsid2, .data$year, .data$duplicated) %>%
+      # keep lowest start month
+      filter(.data$stmon == min(.data$stmon)) %>%
+      arrange(.data$ccode1, .data$obsid1, .data$ccode2, .data$obsid2, .data$year) %>%
+      # practice safe group_by()
+      ungroup() %>%
+      select(-.data$duplicated) -> data
+
+    data[ , c('styear', 'stmon', 'settle', 'fatality', 'mindur', 'maxdur', 'hiact', 'hostlev', 'recip', 'outcome')] <- list(NULL)
+
+
+    attr(data, "ps_data_type") <- attr_ps_data_type
+    attr(data, "ps_system") <-  attr_ps_system
+    attr(data, "ps_conflict_type") <-  attr_ps_conflict_type
 
 
   } else  {